// src/app/modules/issue/interfaces/application.interface.ts
export interface Application {
  cloc_BIK: string;
  cloc_TransferAccount: string;
  cloc_NameClient: string;
  cloc_Address: string;
  cloc_INN: string;
  cloc_Account: string;
  cloc_PayerNameClient: string;
  cloc_PayerINN: string;
  cloc_PayerAccount: string;
  cloc_Amount: string;
  cloc_NDS: string;
  cloc_ValidityPeriod: string;
  cloc_PayerBank: string;
  cloc_PayerBIK: string;
  cloc_Purpose: string;
  cloc_NumberDateContract: string;
  cloc_DocList: string;
}